+++ src/main/java/com/minsproject/league/repository/querydsl/TeamsCustomRepository.java
@@ -1,0 +1,10 @@
+package com.minsproject.league.repository.querydsl;
+
+import com.minsproject.league.entity.Teams;
+
+import java.util.List;
+
+public interface TeamsCustomRepository {
+
+    List<Teams> findByTeamIdGreaterThanOffsetId(Long offsetId);
+}
